+++ src/backendGcp.go
@@ -1386,12 +1386,13 @@ func (d *backendGcp) GetNodeIpMap(name string, internalIPs bool) (map[int]string
 	return nlist, nil
 }
 
-func (d *backendGcp) ClusterListFull(isJson bool, owner string, pager bool, isPretty bool, sort []string) (string, error) {
+func (d *backendGcp) ClusterListFull(isJson bool, owner string, pager bool, isPretty bool, sort []string, renderer string) (string, error) {
 	a.opts.Inventory.List.Json = isJson
 	a.opts.Inventory.List.Owner = owner
 	a.opts.Inventory.List.Pager = pager
 	a.opts.Inventory.List.JsonPretty = isPretty
 	a.opts.Inventory.List.SortBy = sort
+	a.opts.Inventory.List.RenderType = renderer
 	return "", a.opts.Inventory.List.run(d.server, d.client, false, false, false)
 }
 
@@ -2055,11 +2056,12 @@ func (d *backendGcp) VacuumTemplates() error {
 	return d.vacuum(nil)
 }
 
-func (d *backendGcp) TemplateListFull(isJson bool, pager bool, isPretty bool, sort []string) (string, error) {
+func (d *backendGcp) TemplateListFull(isJson bool, pager bool, isPretty bool, sort []string, renderer string) (string, error) {
 	a.opts.Inventory.List.Json = isJson
 	a.opts.Inventory.List.Pager = pager
 	a.opts.Inventory.List.JsonPretty = isPretty
 	a.opts.Inventory.List.SortBy = sort
+	a.opts.Inventory.List.RenderType = renderer
 	return "", a.opts.Inventory.List.run(false, false, true, false, false)
 }